Move files to policy directory
