+++ lib/policies/msRestUserAgentPolicy.browser.ts
@@ -1,6 +1,11 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
+/*
+ * NOTE: When moving this file, please update "browser" section in package.json
+ * and "plugins" section in webpack.testconfig.ts.
+ */
+
 import { TelemetryInfo } from "./userAgentPolicy";
 
 interface NavigatorEx extends Navigator {
