+++ lib/policies/userAgentPolicy.ts
@@ -1,12 +1,12 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-import { BaseRequestPolicy, RequestPolicy, RequestPolicyOptions, RequestPolicyFactory } from "../requestPolicy";
-import { WebResource, HttpHeaders, HttpOperationResponse } from "../../msRest";
-import { Constants } from "../../util/constants";
+import { BaseRequestPolicy, RequestPolicy, RequestPolicyOptions, RequestPolicyFactory } from "./requestPolicy";
+import { WebResource, HttpHeaders, HttpOperationResponse } from "../msRest";
+import { Constants } from "../util/constants";
 import { getPlatformSpecificData, getDefaultUserAgentKey } from "./msRestUserAgentPolicy";
 
-export type TelemetryInfo = { key: string; value?: string };
+export type TelemetryInfo = { key?: string; value?: string };
 
 function getRuntimeInfo(): TelemetryInfo[] {
     const sdkSignature = {
@@ -35,9 +35,9 @@ function getDefaultHeaderValue(): string {
     return userAgent;
 }
 
-export function userAgentPolicy(headerKey?: string, headerValue?: string): RequestPolicyFactory {
-    const key: string = headerKey || getDefaultUserAgentKey();
-    const value: string = headerValue || getDefaultHeaderValue();
+export function userAgentPolicy(userAgentData?: TelemetryInfo): RequestPolicyFactory {
+    const key: string = (userAgentData && userAgentData.key) || getDefaultUserAgentKey();
+    const value: string = (userAgentData && userAgentData.value) || getDefaultHeaderValue();
 
     return {
         create: (nextPolicy: RequestPolicy, options: RequestPolicyOptions) => {
